src/test/javascript/spec/app/entities/course-plan/course-plan.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { CoursePlanService } from 'app/entities/course-plan/course-plan.service';
import { ICoursePlan, CoursePlan } from 'app/shared/model/course-plan.model';
import { Month } from 'app/shared/model/enumerations/month.model';

describe('Service Tests', () => {
  describe('CoursePlan Service', () => {
    let injector: TestBed;
    let service: CoursePlanService;
    let httpMock: HttpTestingController;
    let elemDefault: ICoursePlan;
    let expectedResult: ICoursePlan | ICoursePlan[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(CoursePlanService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new CoursePlan(0, Month.JANUARY, 'AAAAAAA', 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a CoursePlan', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new CoursePlan()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a CoursePlan', () => {
        const returnedFromService = Object.assign(
          {
            month: 'BBBBBB',
            technology: 'BBBBBB',
            extraPrice: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of CoursePlan', () => {
        const returnedFromService = Object.assign(
          {
            month: 'BBBBBB',
            technology: 'BBBBBB',
            extraPrice: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a CoursePlan', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
